Refactor address importer for fewer commands
